(gabac) Update type for numContexts - to align with state_vars::numCtxTotal

+++ src/genie/entropy/gabac/context-tables.cc
@@ -12,7 +12,7 @@ namespace entropy {
 namespace gabac {
 namespace contexttables {
 
-std::vector<ContextModel> buildContextTable(unsigned int numContexts) {
+std::vector<ContextModel> buildContextTable(unsigned long numContexts) {
     return std::vector<ContextModel>(numContexts);
 }
