+++ src/genie/entropy/gabac/context-tables.h
@@ -16,7 +16,7 @@ namespace entropy {
 namespace gabac {
 namespace contexttables {
 
-std::vector<ContextModel> buildContextTable(unsigned int numContexts);
+std::vector<ContextModel> buildContextTable(unsigned long numContexts);
 
 }  // namespace contexttables
 }  // namespace gabac
